Cloud_Functions/purchase-service/requirements.txt
Flask==2.3.2
firebase-admin==6.1.0
Flask-Cors==3.0.10
google-cloud-firestore==2.11.0
google-cloud-pubsub==2.17.1
google-cloud-logging==3.6.0